conf-gen/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "conf-gen"
version = "0.1.0"
description = "Configuration generator for Clash, Quantumult-X, and sing-box proxy clients"
requires-python = ">=3.12"
dependencies = [
    "common",
    "emoji>=2.0.0",
    "packaging>=23.0",
    "pycountry>=22.0.0",
    "pytz>=2023.0",
    "PyYAML>=6.0",
    "requests>=2.28.0",
]

[project.scripts]
conf-gen = "conf_gen._cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
conf_gen = ["py.typed"]

[tool.black]
line-length = 99

[tool.isort]
profile = "black"
line_length = 99

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
no_implicit_optional = true
